requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
Flask-SocketIO==5.3.6
python-barcode==0.15.1
Pillow==10.2.0
simple-websocket==1.0.0
